[NCL-4717]: Rework scheduler classes for deletaion of temporary builds to use interface Remove default constructor creating confussion
